+++ Apizr/Src/Apizr/Configuring/Request/ApizrRequestOptionsBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Threading;
@@ -143,6 +144,47 @@ public class ApizrRequestOptionsBuilder : IApizrRequestOptionsBuilder, IApizrInt
         else
         {
             options.ContextOptionsBuilder += contextOptionsBuilder.Invoke;
+        }
+
+        return this;
+    }
+
+    /// <inheritdoc />
+    public IApizrRequestOptionsBuilder WithLoggedHeadersRedactionNames(IEnumerable<string> redactedLoggedHeaderNames,
+        ApizrDuplicateStrategy strategy = ApizrDuplicateStrategy.Add)
+    {
+        var sensitiveHeaders = new HashSet<string>(redactedLoggedHeaderNames, StringComparer.OrdinalIgnoreCase);
+
+        return WithLoggedHeadersRedactionRule(header => sensitiveHeaders.Contains(header), strategy);
+    }
+
+    /// <inheritdoc />
+    public IApizrRequestOptionsBuilder WithLoggedHeadersRedactionRule(Func<string, bool> shouldRedactHeaderValue,
+        ApizrDuplicateStrategy strategy = ApizrDuplicateStrategy.Add)
+    {
+        switch (strategy)
+        {
+            case ApizrDuplicateStrategy.Ignore:
+                Options.ShouldRedactHeaderValue ??= shouldRedactHeaderValue;
+                break;
+            case ApizrDuplicateStrategy.Add:
+            case ApizrDuplicateStrategy.Merge:
+                if (Options.ShouldRedactHeaderValue == null)
+                {
+                    Options.ShouldRedactHeaderValue = shouldRedactHeaderValue;
+                }
+                else
+                {
+                    var previous = Options.ShouldRedactHeaderValue;
+                    Options.ShouldRedactHeaderValue = header => previous(header) || shouldRedactHeaderValue(header);
+                }
+
+                break;
+            case ApizrDuplicateStrategy.Replace:
+                Options.ShouldRedactHeaderValue = shouldRedactHeaderValue;
+                break;
+            default:
+                throw new ArgumentOutOfRangeException(nameof(strategy), strategy, null);
         }
 
         return this;