+++ Apizr/Src/Apizr/Configuring/Shared/ApizrGlobalSharedOptionsBase.cs
@@ -24,6 +24,7 @@ namespace Apizr.Configuring.Shared
             OperationTimeout = sharedOptions?.OperationTimeout;
             RequestTimeout = sharedOptions?.RequestTimeout;
             Headers = sharedOptions?.Headers ?? new List<string>();
+            ShouldRedactHeaderValue = sharedOptions?.ShouldRedactHeaderValue;
             _contextOptionsBuilder = sharedOptions?.ContextOptionsBuilder;
             _resiliencePropertiesFactories = sharedOptions?.ResiliencePropertiesFactories?.ToDictionary(kpv => kpv.Key, kpv => kpv.Value) ?? 
                                     new Dictionary<string, Func<object>>();
@@ -67,6 +68,9 @@ namespace Apizr.Configuring.Shared
 
         /// <inheritdoc />
         public TimeSpan? RequestTimeout { get; internal set; }
+
+        /// <inheritdoc />
+        public Func<string, bool> ShouldRedactHeaderValue { get; internal set; }
 
         private Action<IApizrResilienceContextOptionsBuilder> _contextOptionsBuilder;
         /// <inheritdoc />