+++ Apizr/Tests/Apizr.Tests/ApizrTests.cs
@@ -797,10 +797,13 @@ namespace Apizr.Tests
                     .WithLogging()
                     .WithBaseAddress("https://reqres.in/api")
                     .WithHeaders("testKey2: testValue2.2", "testKey3: testValue3.1")
+                    .WithLoggedHeadersRedactionNames(new[]{ "testKey2" })
                     .AddDelegatingHandler(watcher));
 
             // Shortcut
-            await apizrTransferManager.ExecuteAsync((opt, api) => api.GetUsersAsync(opt), options => options.WithHeaders("testKey3: testValue3.2", "testKey4: testValue4"));
+            await apizrTransferManager.ExecuteAsync((opt, api) => api.GetUsersAsync(opt), options => 
+                options.WithHeaders("testKey3: testValue3.2", "testKey4: testValue4")
+                    .WithLoggedHeadersRedactionRule(header => header == "testKey3"));
             watcher.Headers.Should().NotBeNull();
             watcher.Headers.Should().ContainKeys("testKey1", "testKey2", "testKey3", "testKey4");
             watcher.Headers.GetValues("testKey1").Should().HaveCount(1).And.Contain("testValue1"); // Set by attribute
@@ -890,7 +893,7 @@ namespace Apizr.Tests
             ex.WithInnerException<OperationCanceledException>();
         }
 
-        [Fact] // todo: iOS fix => TimeoutException
+        [Fact]
         public async Task Cancelling_A_Post_Request_Should_Throw_An_OperationCanceledException()
         {
             var manager = ApizrBuilder.Current.CreateManagerFor<IHttpBinService>(options =>