+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [New/Breaking][Polly] Now **supporting only the brand new Polly v8+ Resilience Strategies/Pipelines/Registry** instead of former Polly v7- Policies
 - [New][HttpTracer] Now we can **filter out unwanted http message parts** with the brand new IgnoreMessageParts fluent option
 - [New][HttpTracer] Now we can **pick RequestAllButBody, HeadersOnly or AllButBody http message parts** when configuring logging options fluently or using attribute
+- [New][HttpTracer] Now we can **redact any header sensitive values from logs** thanks to `WithLoggedHeadersRedactionNames` or `WithLoggedHeadersRedactionRule` fluent options
 - [Breaking][Removed] **Previsoulsy deprecated extensions methods for backward compatibility have been removed** as they were not suitable anymore
 
 ### Apizr.Integrations.MediatR